feat(volo-grpc): change code after review
